requirements.txt
streamlit
PyPDF2==3.0.1
google-generativeai==0.3.0
python-dotenv==1.0.0
streamlit_extras
python-docx

